try to use wlroots from source
the version of wlroots in the ubuntu is quite what i need it seems. i
want to be able to play with compilation settings etc.

so changed to wlroots via a meson subproject.

but this mean i pullled in the latest version of wlroots, which
requires a newer version of meson to compile it.

this is newer than the version of meson supplied in the ubuntu
packages.

so i've followed the instructions on mesons website to install from
pip instead:
  https://mesonbuild.com/Quick-guide.html#requirements

fingers-crossed this'll work (docker build running atm)